move existing export to legacy file naming, update export flag
