+++ packages/compass-import-export/src/components/legacy/export-modal.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { expect } from 'chai';
 
-import store from '../stores/export-store';
+import store from '../../stores/legacy-export-store';
 import ExportModal from './export-modal';
 
 function renderModal() {
