+++ packages/compass-import-export/src/components/legacy/progress-bar/progress-bar.module.less
