+++ packages/compass-import-export/src/export-plugin.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import ExportModal from './components/export-modal';
-import exportStore from './stores/export-store';
+import ExportModal from './components/legacy/export-modal';
+import exportStore from './stores/legacy-export-store';
 
 function ExportPlugin() {
   return (
