+++ packages/compass-preferences-model/src/preferences.ts
@@ -17,7 +17,7 @@ export type FeatureFlags = {
   showDevFeatureFlags?: boolean;
   lgDarkmode?: boolean;
   debugUseCsfleSchemaMap?: boolean;
-  useNewExportBackend?: boolean;
+  useNewExport?: boolean;
   useStageWizard?: boolean;
 };
 
@@ -259,7 +259,7 @@ const featureFlagsProps: Required<{
    * Feature flag for enabling the use of the new backend api for
    * exporting documents. Epic: COMPASS-5576
    */
-  useNewExportBackend: {
+  useNewExport: {
     type: 'boolean',
     required: false,
     default: false,
@@ -267,8 +267,8 @@ const featureFlagsProps: Required<{
     cli: true,
     global: true,
     description: {
-      short: 'New Export Backend',
-      long: 'Use the new backend api for exporting documents.',
+      short: 'New Export',
+      long: 'Use the new export backend and views when exporting documents.',
     },
   },
